Update copyright line in license
Changed the name to "openMSX Team". This is not a legal entity, only
a placeholder; check the Git history to find the author who is the
actual copyright holder.

I strongly doubt anyone is going to bother to update the year in this
file, so I removed it. If needed, the change dates can be found in
the Git history.